+++ draft-packages/collapsible/package.json
@@ -33,6 +33,7 @@
   "license": "MIT",
   "dependencies": {
     "@kaizen/component-library": "^12.7.0",
+    "@kaizen/typography": "1.1.0",
     "classnames": "^2.3.1",
     "react-animate-height": "^2.0.23"
   },
