+++ draft-packages/divider/docs/Divider.stories.tsx
@@ -2,7 +2,8 @@ import React from "react"
 import { Divider } from "@kaizen/draft-divider"
 import { Story } from "@storybook/react"
 import { Card } from "@kaizen/draft-card"
-import { Box, Heading, Paragraph } from "@kaizen/component-library"
+import { Box } from "@kaizen/component-library"
+import { Heading, Paragraph } from "@kaizen/typography"
 import { withDesign } from "storybook-addon-designs"
 import { figmaEmbed } from "../../../storybook/helpers"
 import { CATEGORIES } from "../../../storybook/constants"
