+++ draft-packages/form/docs/TextField.stories.tsx
@@ -1,6 +1,6 @@
 import React from "react"
 import { withDesign } from "storybook-addon-designs"
-import { Heading } from "@kaizen/component-library"
+import { Heading } from "@kaizen/typography"
 import { TextField } from "@kaizen/draft-form"
 import dateIcon from "@kaizen/component-library/icons/date-start.icon.svg"
 import { figmaEmbed } from "../../../storybook/helpers"
