+++ draft-packages/guidance-block/package.json
@@ -33,6 +33,7 @@
   "license": "MIT",
   "dependencies": {
     "@kaizen/component-library": "^12.7.0",
+    "@kaizen/typography": "^1.1.0",
     "@kaizen/draft-button": "^5.6.13",
     "@kaizen/draft-illustration": "^3.6.13",
     "@kaizen/draft-tooltip": "^3.7.10",
