+++ draft-packages/modal/KaizenDraft/Modal/Presets/InputEditModal.tsx
@@ -1,6 +1,6 @@
 import classnames from "classnames"
 import * as React from "react"
-import { Heading } from "@kaizen/component-library"
+import { Heading } from "@kaizen/typography"
 import { ButtonProps } from "@kaizen/draft-button"
 import {
   GenericModal,
