+++ draft-packages/modal/package.json
@@ -35,6 +35,7 @@
     "@headlessui/react": "^1.4.3",
     "@kaizen/component-library": "^12.7.0",
     "@kaizen/deprecated-component-library-helpers": "^2.5.2",
+    "@kaizen/typography": "1.1.0",
     "@kaizen/draft-button": "^5.6.13",
     "@kaizen/draft-divider": "^1.9.3",
     "@kaizen/draft-form": "^5.6.12",
