+++ draft-packages/popover/package.json
@@ -33,6 +33,7 @@
   "license": "MIT",
   "dependencies": {
     "@kaizen/component-library": "^12.7.0",
+    "@kaizen/typography": "1.1.0",
     "@kaizen/deprecated-component-library-helpers": "^2.5.2",
     "@popperjs/core": "^2.11.0",
     "@types/classnames": "^2.3.0",
