+++ draft-packages/table/package.json
@@ -33,6 +33,7 @@
   "license": "MIT",
   "dependencies": {
     "@kaizen/component-library": "^12.7.0",
+    "@kaizen/typography": "1.1.0",
     "@kaizen/draft-form": "^5.6.12",
     "@kaizen/draft-tooltip": "^3.7.10",
     "@types/classnames": "^2.3.0",
