+++ draft-packages/title-block-zen/package.json
@@ -33,6 +33,7 @@
   "license": "MIT",
   "dependencies": {
     "@kaizen/component-library": "^12.7.0",
+    "@kaizen/typography": "1.1.0",
     "@kaizen/draft-avatar": "^2.3.15",
     "@kaizen/draft-badge": "^1.10.4",
     "@kaizen/draft-button": "^5.6.13",
