+++ draft-packages/tooltip/docs/Tooltip.stories.tsx
@@ -3,7 +3,8 @@ import React from "react"
 import informationIcon from "@kaizen/component-library/icons/information-white.icon.svg"
 import meatballsIcon from "@kaizen/component-library/icons/meatballs.icon.svg"
 import { Tag } from "@kaizen/draft-tag"
-import { Icon, Paragraph, Heading } from "@kaizen/component-library"
+import { Icon } from "@kaizen/component-library"
+import { Paragraph, Heading } from "@kaizen/typography"
 import { withDesign } from "storybook-addon-designs"
 import { Button, IconButton } from "@kaizen/button"
 import { Tooltip } from "@kaizen/draft-tooltip"
