+++ packages/progress-bar/package.json
@@ -37,6 +37,7 @@
   },
   "dependencies": {
     "@kaizen/component-library": "^12.7.0",
+    "@kaizen/typography": "1.1.0",
     "classnames": "^2.3.1"
   }
 }
