+++ storybook/components/StoryWrapper/components/StoryRowHeader/StoryRowHeader.tsx
@@ -1,6 +1,6 @@
 import React from "react"
 import classNames from "classnames"
-import { Heading } from "../../../../../packages/component-library"
+import { Heading } from "../../../../../packages/typography"
 import styles from "../../StoryWrapper.scss"
 
 export interface StoryRowHeaderProps {
